+++ scripts/importer/mtasks/asiago.py
@@ -14,8 +14,8 @@ from .. funcs import add_event, add_source, add_quantity, \
 from ... utils import pbar, is_number
 
 
-def do_asiago_photo(events, args, tasks):
-    current_task = 'asiago: photometry'
+def do_asiago_photo(events, args, tasks, task_obj):
+    current_task = task_obj.current_task(args)
     # response = urllib.request.urlopen('http://graspa.oapd.inaf.it/cgi-bin/sncat.php')
     path = os.path.abspath(os.path.join(PATH.REPO_EXTERNAL, 'asiago-cat.php'))
     response = urllib.request.urlopen('file://' + path)
@@ -111,9 +111,9 @@ def do_asiago_photo(events, args, tasks):
     return events
 
 
-def do_asiago_spectra(events, args, tasks):
-    current_task = 'asiago: spectra'
-    html = load_cached_url(args, 'http://sngroup.oapd.inaf.it./cgi-bin/output_class.cgi?sn=1990',
+def do_asiago_spectra(events, args, tasks, task_obj):
+    current_task = task_obj.current_task(args)
+    html = load_cached_url(args, current_task, 'http://sngroup.oapd.inaf.it./cgi-bin/output_class.cgi?sn=1990',
                            os.path.join(PATH.REPO_EXTERNAL_SPECTRA, 'Asiago/spectra.html'))
     if not html:
         return events