+++ scripts/importer/mtasks/suspect.py
@@ -19,8 +19,8 @@ from .. funcs import add_event, add_photometry, add_source, add_quantity, add_sp
 from ... utils import get_sig_digits, is_number, pbar, pbar_strings, pretty_num
 
 
-def do_suspect_photo(events, args, tasks):
-    current_task = 'Suspect: Photometry'
+def do_suspect_photo(events, args, tasks, task_obj):
+    current_task = task_obj.current_task(args)
     with open(os.path.join(PATH.REPO_EXTERNAL, 'suspectreferences.csv'), 'r') as f:
         tsvin = csv.reader(f, delimiter=',', skipinitialspace=True)
         suspectrefdict = {}
@@ -99,8 +99,8 @@ def do_suspect_photo(events, args, tasks):
     return events
 
 
-def do_suspect_spectra(events, args, tasks):
-    current_task = 'Suspect: Spectra'
+def do_suspect_spectra(events, args, tasks, task_obj):
+    current_task = task_obj.current_task(args)
     with open(os.path.join(PATH.REPO_EXTERNAL_SPECTRA, 'Suspect/sources.json'), 'r') as f:
         sourcedict = json.loads(f.read())